refactored - extracted `<BoardComponent/>`
