+++ src/minesweeper/reactHooks/PlainReactHookMinesweeperApp.tsx
@@ -7,12 +7,10 @@ import produce from "immer";
 import {blankBoardData, initializedBoardData} from "../utils";
 import {BoardData, BoardDimension, CellCoordinates} from "../types";
 import GameConfigDialog from "./GameConfigDialog";
+import BoardComponent from "./BoardComponent";
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
-        table: {
-            minWidth: 650,
-        },
         gameAppContainer: {
             "minHeight": "85vh",
             "display": "flex",
@@ -25,19 +23,6 @@ const useStyles = makeStyles((theme: Theme) =>
             borderRadius: theme.spacing(1),
             backgroundColor: theme.palette.background.paper,
         },
-        boardContainer: {
-            display: "block",
-            width: "max-content",
-        },
-        cellButtonContainer: {
-            margin: theme.spacing(0.25),
-            display: "inline-block",
-        },
-        cellButton: {
-            height: theme.spacing(8),
-            width: theme.spacing(8),
-            minWidth: "unset",
-        }
     }),
 );
 
@@ -107,39 +92,6 @@ function PlainReactHookMinesweeperApp() {
             }
         }))
     }
-}
-
-function BoardComponent(
-    {data, isPlaying, handleClick}:
-        { data: BoardData, isPlaying: boolean, handleClick: (coor: CellCoordinates) => void }
-) {
-    const classes = useStyles();
-
-    return (
-        <div className={classes.boardContainer}>
-            {data.map((rowData, row) => (
-                <div key={`row_${row}`}>
-                    {
-                        rowData.map((cellData, col) => (
-                            <div
-                                key={`cell_${row * rowData.length + col}`}
-                                className={classes.cellButtonContainer}
-                            >
-                                <Button
-                                    variant="contained"
-                                    color="default"
-                                    className={classes.cellButton}
-                                    children={cellData.isBomb ? "B" : cellData.count !== 0 ? cellData.count : ""}
-                                    disabled={cellData.isOpen || !isPlaying}
-                                    onClick={() => handleClick({row, col})}
-                                />
-                            </div>
-                        ))
-                    }
-                </div>
-            ))}
-        </div>
-    )
 }
 
 export default PlainReactHookMinesweeperApp;
